+++ zaifbot/modules/auto_trade/__init__.py
@@ -1,0 +1,3 @@
+from .loss_cut import LossCut
+
+__all__ = ['LossCut']
