--- zaifbot/modules/auto_trade/process/__init__.py
@@ -1,3 +1,0 @@
-from .loss_cut import LossCut
-
-__all__ = ['LossCut']
